Fixed background-width scrolling issue
Fixed an issue with chrome rendering the background-attachment:fixed
properly very sluggishly.  Found a attached the fixed bg to the
element’s ::before tag and use the will-change:transform property.  The
background of the image is set in a separate class as well.

Codepen: http://codepen.io/of6/pen/dorPRX